Moved initialize to a general folder
